Seperate client implementation
